Fixes trac:2702: Move new ROMIO to V1.5
r24264
Romio Refresh from mpich2-1.3.1. Work by Pascal Deveze, tested through bitbucket by Jeff Squyres (https://bitbucket.org/devezep/new-romio-for-openmpi).

r24272
Fixes bug introduced by r24264 : make Romio autogen.sh executable.

r24285
Add missing line continuation character to prevent a Makefile syntax
error

r24292
Fix all the problems with "make distcheck" caused by the new ROMIO import so that we can finally get a trunk nightly tarball\!

This commit was SVN r24335.

The following SVN revision numbers were found above:
  r24264 --> open-mpi/ompi@0e921bba7fee80795269bc22d9af69294eebb32d
  r24272 --> open-mpi/ompi@46b711e164be6be6706dc19ffbf9c5b77ad683fa
  r24285 --> open-mpi/ompi@89fb26eb1c2cbc76044a64f06a52b84a374e10dd
  r24292 --> open-mpi/ompi@30e164e2469351eccb7fd0a88f681bc485ab46d0

The following Trac tickets were found above:
  Ticket 2702 --> https://svn.open-mpi.org/trac/ompi/ticket/2702

+++ ompi/mca/io/romio/romio/adio/ad_testfs/ad_testfs_write.c
@@ -23,7 +23,7 @@ void ADIOI_TESTFS_WriteContig(ADIO_File fd, void *buf, int count,
 	    nprocs, fd->filename);
     FPRINTF(stdout, "[%d/%d]    writing (buf = %p, loc = %lld, sz = %lld)\n",
 	    myrank, nprocs, buf, (long long) offset, 
-	    (long long) datatype_size * count);
+	    (long long)datatype_size * (long long)count);
 
     if (file_ptr_type != ADIO_EXPLICIT_OFFSET)
     {
